+++ apps/ergo4all/lib/app.dart
@@ -4,12 +4,14 @@ import 'package:ergo4all/analysis/live/screen.dart';
 import 'package:ergo4all/common/routes.dart';
 import 'package:ergo4all/gen/i18n/app_localizations.dart';
 import 'package:ergo4all/home/screen.dart';
+import 'package:ergo4all/imprint_screen.dart';
 import 'package:ergo4all/language/screen.dart';
 import 'package:ergo4all/onboarding/expert_intro_screen.dart';
 import 'package:ergo4all/onboarding/non_expert_intro_screen.dart';
 import 'package:ergo4all/onboarding/pre_intro_screen.dart';
 import 'package:ergo4all/onboarding/pre_user_creator_screen.dart';
 import 'package:ergo4all/onboarding/user_creator_screen.dart';
+import 'package:ergo4all/privacy_screen.dart';
 import 'package:ergo4all/results/detail/screen.dart';
 import 'package:ergo4all/results/overview/screen.dart';
 import 'package:ergo4all/route_leave_observer.dart';
@@ -74,6 +76,8 @@ class _Ergo4AllAppState extends State<Ergo4AllApp> {
         Routes.welcome.path: (context) => const WelcomeScreen(),
         Routes.tipChoice.path: (context) => const TipChoiceScreen(),
         Routes.tipDetail.path: (context) => const TipDetailScreen(),
+        Routes.imprint.path: (_) => const ImprintScreen(),
+        Routes.privacy.path: (_) => const PrivacyScreen(),
       },
       navigatorObservers: [
         RouteLeaveObserver(