+++ apps/ergo4all/lib/home/menu_dialog.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:common_ui/theme/colors.dart';
 import 'package:common_ui/theme/styles.dart';
 import 'package:ergo4all/common/routes.dart';
@@ -30,6 +32,11 @@ Future<void> showHomeMenuDialog(BuildContext context) {
     Navigator.of(context).pushReplacementNamed(Routes.home.path);
   }
 
+  void goToPrivacy() {
+    final navigator = Navigator.of(context)..pop();
+    unawaited(navigator.pushNamed(Routes.privacy.path));
+  }
+
   return showDialog<void>(
     context: context,
     builder: (context) => SimpleDialog(
@@ -48,14 +55,11 @@ Future<void> showHomeMenuDialog(BuildContext context) {
           goToLanguage,
           const Key('button-lang'),
         ),
-        _makeOptionButton(localizations.menu_imprint_label, () {
-          Navigator.of(context).pop();
-          showNotImplementedSnackbar(context);
+        _makeOptionButton(localizations.imprint, () {
+          final navigator = Navigator.of(context)..pop();
+          unawaited(navigator.pushNamed(Routes.imprint.path));
         }),
-        _makeOptionButton(localizations.menu_privacy_label, () {
-          Navigator.of(context).pop();
-          showNotImplementedSnackbar(context);
-        }),
+        _makeOptionButton(localizations.menu_privacy_label, goToPrivacy),
         _makeOptionButton(localizations.menu_delete_data_label, () {
           Navigator.of(context).pop();
           showNotImplementedSnackbar(context);