+++ apps/ergo4all/lib/home/screen.dart
@@ -42,6 +42,18 @@ class HomeScreen extends HookWidget {
     );
 
     return Scaffold(
+      appBar: RedCircleAppBar(
+        titleText: localizations.home_title,
+        menuButton: IconButton(
+          key: const Key('burger'),
+          onPressed: () {
+            showHomeMenuDialog(context);
+          },
+          icon: const Icon(Icons.menu),
+          color: white,
+          iconSize: 48,
+        ),
+      ),
       body: Column(
         mainAxisSize: MainAxisSize.min,
         children: [