+++ packages/common_ui/lib/widgets/red_circle_top_bar.dart
@@ -1,15 +1,21 @@
 import 'package:common_ui/theme/colors.dart';
+import 'package:common_ui/theme/spacing.dart';
 import 'package:common_ui/theme/styles.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 
+const double _appBarHeight = 200;
+
 /// The red circle element at the top of a lot of pages. See the
 /// [Figma design for the Home screen](https://www.figma.com/design/ToLbPoKLTVnyU33ItvwLuH/Ergo4All-draft-for-new-design?node-id=123-520&t=Xp1FzTpSjlknOlQY-4)
 /// for an example.
-class RedCircleTopBar extends StatelessWidget {
+///
+/// You can use it for [Scaffold#appBar].
+class RedCircleAppBar extends StatelessWidget implements PreferredSizeWidget {
   /// Construct a circle bar.
-  const RedCircleTopBar({
+  const RedCircleAppBar({
     required this.titleText,
+    this.withBackButton = false,
     this.menuButton,
     super.key,
   });
@@ -17,34 +23,49 @@ class RedCircleTopBar extends StatelessWidget {
   /// The title text displayed on the circle.
   final String titleText;
 
+  /// Whether to include a back-button in the bar.
+  final bool withBackButton;
+
   /// An optional (pressable) widget which is displayed on the top right of the
   /// bar.
   final Widget? menuButton;
 
   @override
   Widget build(BuildContext context) {
-    return Stack(
-      alignment: AlignmentDirectional.center,
-      children: [
-        FittedBox(
-          fit: BoxFit.cover,
-          child: SvgPicture.asset(
-            'assets/images/top_circle.svg',
+    return SizedBox(
+      height: _appBarHeight,
+      child: Stack(
+        alignment: AlignmentDirectional.center,
+        children: [
+          SvgPicture.asset(
+            fit: BoxFit.fill,
             package: 'common_ui',
+            'assets/images/top_circle.svg',
           ),
-        ),
-        if (menuButton != null && menuButton?.key?.toString() != "[<'burger'>]")
-          Positioned(top: 70, left: 20, child: menuButton!),
-        if (menuButton != null && menuButton?.key?.toString() == "[<'burger'>]")
-          Positioned(top: 70, right: 40, child: menuButton!),
-        Positioned(
-          top: 120,
-          child: Text(
-            titleText,
-            style: h1Style.copyWith(color: white),
+          if (withBackButton)
+            const Positioned(
+              top: 50,
+              left: mediumSpace,
+              child: BackButton(color: white),
+            ),
+          if (menuButton != null)
+            Positioned(
+              top: 50,
+              right: mediumSpace,
+              child: menuButton!,
+            ),
+          Positioned(
+            top: 100,
+            child: Text(
+              titleText,
+              style: h1Style.copyWith(color: white),
+            ),
           ),
-        ),
-      ],
+        ],
+      ),
     );
   }
+
+  @override
+  Size get preferredSize => const Size.fromHeight(_appBarHeight);
 }